Read json from the bagit package.
We've got rid of the three CSV files and replaced it with a single bagit
json. This change reads from that file and writes it to Dynamo.

It's more or less a copy of the json in the bagit package into Dynamo.
We're calculating the parent path from the ID and adding in the
checksums from the bag metadata and the file extensions and adding them
to the dynamo json.

This is all done with json objects and string keys. It's a bit fiddly
but as there'll be fields that we won't know in advance, we can't use a
case class.